+++ src/shims-worker.d.ts
@@ -1,0 +1,11 @@
+declare module '*?worker' {
+  export default class WebpackWorker extends Worker {
+    constructor();
+  }
+}
+
+declare module '*?worker&inline' {
+  export default class WebpackWorker extends Worker {
+    constructor();
+  }
+}